Redesign tests and pages structure
