+++ cypress/e2e/fleet.cy.js
@@ -1,22 +1,30 @@
+import { fleetsPage } from '../views/fleetsPage'
+import { repositoriesPage } from '../views/repositoriesPage'
+
 describe('Fleet Management', () => {
-    beforeEach(() => {
-        cy.login(`${Cypress.env('host')}`, `${Cypress.env('auth')}`, `${Cypress.env('username')}`, `${Cypress.env('password')}`)
-    })
-    it('Should create a fleet', () => {
-        cy.createFleet(`${Cypress.env('image')}`)
-    })
-  /*   it('Should edit a fleet', () => {
-         cy.editFleet(`${Cypress.env('image')}`)
-     }) */
-    it('Should delete a fleet', () => {
-         cy.deleteFleet('test-fleet')
-    })
-    it('Should import a fleet', () => {
-        cy.importFleet()
-    })
- /*    it('Should delete an imported fleet', () => {
-        cy.deleteRepository('test-fleet')
-        cy.wait(5000)
-        cy.deleteFleet('basic-nginx-fleet')
-    }) */
+  beforeEach(() => {
+    cy.login(`${Cypress.env('host')}`, `${Cypress.env('auth')}`, `${Cypress.env('username')}`, `${Cypress.env('password')}`)
+  })
+
+  it('Should create a fleet', () => {
+    fleetsPage.createFleet(`${Cypress.env('image')}`)
+  })
+
+  /* it('Should edit a fleet', () => {
+    fleetsPage.editFleet(`${Cypress.env('image')}`)
+  }) */
+
+  it('Should delete a fleet', () => {
+    fleetsPage.deleteFleet('test-fleet')
+  })
+
+  it('Should import a fleet', () => {
+    fleetsPage.importFleet()
+  })
+
+  /* it('Should delete an imported fleet', () => {
+    repositoriesPage.deleteRepository('test-fleet')
+    cy.wait(5000)
+    fleetsPage.deleteFleet('basic-nginx-fleet')
+  }) */
 })
