+++ cypress/e2e/repository.cy.js
@@ -1,15 +1,19 @@
-describe('Repository Management', () => {
-    beforeEach(() => {
-        cy.login(`${Cypress.env('host')}`, `${Cypress.env('auth')}`, `${Cypress.env('username')}`, `${Cypress.env('password')}`)
-    })
-/*     it('Should create a repository', () => {
-        cy.createRepository(`${Cypress.env('repository')}`, `${Cypress.env('revision')}`, `${Cypress.env('resource')}`)
+import { repositoriesPage } from '../views/repositoriesPage'
 
-    }) */
-    it('Should edit a repository', () => {
-        cy.editRepository(`${Cypress.env('repository')}`)
-    })
-/*     it('Should delete a repository', () => { 
-        cy.deleteRepository(`${Cypress.env('repository')}`)
-    }) */
+describe('Repository Management', () => {
+  beforeEach(() => {
+    cy.login(`${Cypress.env('host')}`, `${Cypress.env('auth')}`, `${Cypress.env('username')}`, `${Cypress.env('password')}`)
+  })
+
+  /* it('Should create a repository', () => {
+    repositoriesPage.createRepository(`${Cypress.env('repository')}`, `${Cypress.env('revision')}`, `${Cypress.env('resource')}`)
+  }) */
+
+  it('Should edit a repository', () => {
+    repositoriesPage.editRepository(`${Cypress.env('repository')}`)
+  })
+
+  /* it('Should delete a repository', () => {
+    repositoriesPage.deleteRepository(`${Cypress.env('repository')}`)
+  }) */
 })
